% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimal_places.R
\name{decimal_places}
\alias{decimal_places}
\title{Get decimal places of coordinate values}
\usage{
decimal_places(coord)
}
\arguments{
\item{coord}{A coordinate value in the numeric format of decimal degree}
}
\value{
a numerical value which represent the number of decimal places for the coordiante
Get decimal places of coordinate values
}
\description{
Get decimal places of coordinate values
}
\examples{
decimal_places(12.7000000)
decimal_places(45.67788)

}
