% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_flag_non_west_pacific.R
\name{ec_flag_non_west_pacific}
\alias{ec_flag_non_west_pacific}
\title{flag data points not in east pacific}
\usage{
ec_flag_non_west_pacific(ocean_names, buffer_distance = 50000, ecodata)
}
\arguments{
\item{ocean_names, }{Insert the name of oceans:
"South Pacific Ocean", "North Pacific Ocean", North Atlantic Ocean", "South Atlantic Ocean"}

\item{buffer_distance, }{Its a certain buffer distance to consider if a data point is inland. Beyond this distance data points consider as bad data points. e.g. buffer_distance <- 25000}

\item{ecodata, }{Data table which has latitude and longitude information}
}
\value{
a new column with flagged values, 1 means bad records 0 means good record. Column name: flag_non_region
}
\description{
flag data points not in east pacific
}
\examples{
\dontrun{
ocean_names <- c("North Pacific Ocean", "South Pacific Ocean")
buffer_distance <- 25000
data <- data.frame(
  species = "A",
  decimalLongitude = c(-120, -78, -110),
  decimalLatitude = c(20, 34, 30)
)
data$flag_non_region <- ec_flag_non_west_pacific(ocean_names, buffer_distance, data)
}

}
