% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_geographic_map.R
\name{ec_geographic_map}
\alias{ec_geographic_map}
\title{a map view of data}
\usage{
ec_geographic_map(
  data,
  decimalLatitude = "decimalLatitude",
  decimalLongitude = "decimalLongitude"
)
}
\arguments{
\item{data}{Data table which has information of coordinates (decimalLongitude and decimalLatitude), you don't have to call any flagged column as this function will display all data points of the input data table}

\item{decimalLatitude}{default set to "decimalLatitude"}

\item{decimalLongitude}{default set to "decimalLongitude"}
}
\value{
A map view shows occurrence records.
}
\description{
a map view of data
}
\examples{
data <- data.frame(
  scientificName = "Mexacanthina lugubris",
  decimalLongitude = c(-117, -117.8, -116.9),
  decimalLatitude = c(32.9, 33.5, 31.9),
  temperature_mean = c(12, 13, 14),
  temperature_min = c(9, 6, 10),
  temperature_max = c(14, 16, 18)
)
ec_geographic_map(data)
}
