% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_merge_corrected_coordinates.R
\name{ec_merge_corrected_coordinates}
\alias{ec_merge_corrected_coordinates}
\title{merge the update georeferences back to the main file.}
\usage{
ec_merge_corrected_coordinates(data_corrected, data)
}
\arguments{
\item{data_corrected}{After assigning coordinate values using online georeference tools such as GeoLocate, upload the csv file back to R with the name call data_corrected}

\item{data}{data table which needs to updated with the assign coordiantes}
}
\value{
data table with updated coordinate information
}
\description{
merge the update georeferences back to the main file.
}
\examples{
data <- data.frame(
 species = "A",
 decimalLongitude = c(-120, -119.8, NA, NA),
 decimalLatitude = c(20, 34, NA, NA),
 cleaned_catalog = c("12345", "89888", "LACM8898", "SDNHM6767"),
 locality = c(NA, NA, "Los Angeles, CA", "San Pedro, CA"),
 coordinateUncertaintyInMeters = c(9999, NA, NA, NA)
)
data_corrected <- data.frame(
 species = "A",
 decimalLongitude = c(-120, -119.8, 118, 118.3),
 decimalLatitude = c(20, 34, 33, 32.9),
 cleaned_catalog = c("12345", "89888", "LACM8898", "SDNHM6767"),
 locality = c(NA, NA, "Los Angeles, CA", "San Pedro, CA"),
 coordinateUncertaintyInMeters = c(9999, NA, 5000, 1000)
)


data<- ec_merge_corrected_coordinates(data, data_corrected)

}
