% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_rm_duplicate.R
\name{ec_rm_duplicate}
\alias{ec_rm_duplicate}
\title{remove duplicate records based on catalogNumber and abundance columns.
mandatory fields are catalogNumber, source and abundance}
\usage{
ec_rm_duplicate(data, catalogNumber = "catalogNumber", abundance = "abundance")
}
\arguments{
\item{data}{this is merge data frame which is a output file after running ec_db_merge}

\item{catalogNumber}{this is a mandatory field which consider unique for each occurrence record.}
}
\value{
A data frame which has unique catalog numbers. the output file will have cleaned_catalog field instead of catalogNumber. Also the unique record will be chosen with the abundance value if there is any.
}
\description{
remove duplicate records based on catalogNumber and abundance columns.
mandatory fields are catalogNumber, source and abundance
}
\examples{

db1<- data.frame(species = "A",
                  decimalLongitude = c(-120.2, -117.1, NA, NA),
                  decimalLatitude = c(20.2, 34.1, NA, NA),
                  catalogNumber = c("12345", "89888", "LACM8898", "SDNHM6767"),
                  occurrenceStatus = c("present","","ABSENT", "Present"),
                  basisOfRecord = c("preserved_specimen", "","fossilspecimen", "material_sample"),
                  source = "db1",
                  abundance = c(1,NA,8,23)
                  )

db2<- data.frame(species = "A",
                  decimalLongitude = c(-120.2, -117.1, NA, NA),
                  decimalLatitude = c(20.2, 34.1, NA, NA),
                  catalogNumber = c("123452", "898828", "LACM82898", "SDNHM62767"),
                  occurrenceStatus = c("present","","ABSENT", "Present"),
                  basisOfRecord = c("preserved_specimen", "","fossilspecimen", "material_sample"),
                  source = "db2",
                  abundance = c(1,2,3,19)
                  )
merge_modern_data <- ec_db_merge("modern", db1, db2)
ecodata <- ec_rm_duplicate(merge_modern_data, catalogNumber = "catalogNumber",abundance = "abundance")

}
