% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_var_summary.R
\name{ec_var_summary}
\alias{ec_var_summary}
\title{This function use for getting summary table of spatial and environmental variable for a species habitat suitability}
\usage{
ec_var_summary(data, env_layers)
}
\arguments{
\item{data}{data table after cleaning the records}

\item{env_layers}{an array of col names of enviornmental layers}
}
\value{
a summary table with the mean, min and max values of spatial and environmenal variables
}
\description{
This function use for getting summary table of spatial and environmental variable for a species habitat suitability
}
\examples{
data <- data.frame(
  scientificName = "Mexacanthina lugubris",
  decimalLongitude = c(-117, -117.8, -116.9, -116.5),
  decimalLatitude = c(32.9, 33.5, 31.9, 32.4),
  temperature_mean = c(12, 13, 14, 11),
  temperature_min = c(9, 6, 10, 10),
  temperature_max = c(14, 16, 18, 17)
  )
env_layers <- c("temperature_mean", "temperature_min", "temperature_max")
ec_var_summary(data, env_layers)


}
