% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ecodata}
\alias{ecodata}
\title{dataset1: Documentation of data file - ecodata.rda}
\format{
A data frame with 1115 rows and 19 variables:
\describe{
\item{X}{index}
\item{basisOfRecord}{Type of record (e.g., preserved specimen, fossil)}
\item{occurrenceStatus}{Presence or absence of the organism}
\item{institutionCode}{Code of the institution that holds the record}
\item{verbatimEventDate}{Original recorded date of the event}
\item{scientificName}{Full scientific name of the organism}
\item{individualCount}{Number of individuals observed}
\item{organismQuantity}{Reported quantity of the organism}
\item{abundance}{Calculated or standardized abundance value}
\item{decimalLatitude}{Latitude in decimal degrees}
\item{decimalLongitude}{Longitude in decimal degrees}
\item{coordinateUncertaintyInMeters}{Uncertainty in coordinates (meters)}
\item{locality}{Named place where the occurrence was recorded}
\item{verbatimLocality}{Original text for locality description}
\item{municipality}{Municipality or town of the occurrence}
\item{county}{County where the record was observed}
\item{stateProvince}{State or province name}
\item{country}{Country name}
\item{cleaned_catalog}{Standardized catalog number for de-duplication}
}
}
\source{
\itemize{
\item used rgbif for GBIF, ridigbio for iDigBio, robis for OBIS and rsymbiota for InvertEBase
}
}
\usage{
ecodata
}
\description{
This data file is consider as raw data file after merging and removing duplicate records of all data sources. e.g. this file is an output of occurrence records of mollusc species "Mexacanthina lugubris" with all modern records extracted from GBIF, OBIS, IDIGBIO and InvertEBase
}
\keyword{datasets}
