% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ecodata_corrected}
\alias{ecodata_corrected}
\title{dataset2: Documentation of data file - ecodata_corrected.rda}
\format{
A data frame with 433 rows and 4 variables:
\describe{
\item{cleaned_catalog}{catalog number}
\item{corrected_latitude}{latitude values assigned by GEOLocate}
\item{corrected_longitude}{longitude values assigned by GEOLocate}
\item{corrected_uncertainty}{uncertainty values assigned by GEOLocate}
}
}
\source{
\itemize{
\item this file was created manually after extracting the csv file from GEOLocate online software to assign coordiante and uncertainty values for the records has locality information
}
}
\usage{
ecodata_corrected
}
\description{
This data file created by using GEOLocate tool and we only kept 4 columns. These georeference information will be merge back with the main data file ecodata
}
\keyword{datasets}
