% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ecodata_with_outliers}
\alias{ecodata_with_outliers}
\title{dataset3: Documentation of data file - ecodata_with_outliers.rda}
\format{
A data frame with 713 rows and 35 variables:
\describe{
\item{X}{index}
\item{basisOfRecord}{Type of occurrence record (e.g., preserved specimen, fossil)}
\item{occurrenceStatus}{Indicates presence or absence of the species}
\item{institutionCode}{Code of the institution that provided the record}
\item{verbatimEventDate}{Original text for the event or collection date}
\item{scientificName}{Scientific name of the organism}
\item{individualCount}{Number of individuals recorded}
\item{organismQuantity}{Reported quantity (unit may vary)}
\item{abundance}{Standardized or calculated abundance value}
\item{decimalLatitude}{Latitude in decimal degrees}
\item{decimalLongitude}{Longitude in decimal degrees}
\item{coordinateUncertaintyInMeters}{Spatial uncertainty of coordinates in meters}
\item{locality}{Named location where the record was collected}
\item{verbatimLocality}{Original locality text as provided by the source}
\item{municipality}{Municipality or town of occurrence}
\item{county}{County of occurrence}
\item{stateProvince}{State or province of occurrence}
\item{country}{Country of occurrence}
\item{cleaned_catalog}{Standardized catalog number used for de-duplication}
\item{lat_precision}{Number of decimal places in the latitude coordinate}
\item{lon_precision}{Number of decimal places in the longitude coordinate}
\item{flag_cordinate_precision}{Flag for low coordinate precision}
\item{flag_cc_val}{Flag for invalid or impossible coordinates}
\item{flag_cc_equal}{Flag for identical latitude and longitude (likely erroneous)}
\item{flag_cc_zero}{Flag for coordinates at (0,0)}
\item{flag_cc_cent}{Flag for coordinates placed at a country or region centroid}
\item{flag_cc_gbif}{Flag for coordinates matching GBIF headquarters (artifact)}
\item{flag_cc_inst}{Flag for coordinates matching institution location}
\item{flag_non_region}{Flag for coordinates outside the study region}
\item{outliers}{Flag for outliers based clustering of spatial and env variables}
\item{BO_sstmean}{Mean sea surface temperature from Bio-ORACLE}
\item{BO_sstmax}{Maximum sea surface temperature from Bio-ORACLE}
\item{BO_sstmin}{Minimum sea surface temperature from Bio-ORACLE}
\item{BO_chloro}{Chlorophyll concentration from Bio-ORACLE}
\item{BO_dissox}{Dissolved oxygen level from Bio-ORACLE}
}
}
\source{
\itemize{
\item this file was created manually after extracting the csv file from GEOLocate online software to assign coordiante and uncertainty values for the records has locality information
}
}
\usage{
ecodata_with_outliers
}
\description{
This data file created after running ec_flag_outlier function. It has records with outlier probability
}
\keyword{datasets}
