% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ecodata_x}
\alias{ecodata_x}
\title{dataset5: Documentation of data file - ecodata_x.rda}
\format{
A data frame with 705 rows and 6 variables:
\describe{
\item{species}{species name}
\item{decimalLatitude}{Latitude in decimal degrees}
\item{decimalLongitude}{Longitude in decimal degrees}
\item{temperature_mean_BO}{Mean sea surface temperature from Bio-ORACLE}
\item{temperature_max_BO}{Maximum sea surface temperature from Bio-ORACLE}
\item{temperature_min_BO}{Minimum sea surface temperature from Bio-ORACLE}
}
}
\source{
\itemize{
\item this file has unique coordinate information with unique values of enviornemnt variables
}
}
\usage{
ecodata_x
}
\description{
This data was created to get unique combination of coordinate values to extract env variables from bio-oracle and merge back in main data table - ecodata
}
\keyword{datasets}
