% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{example_sp_invertebase}
\alias{example_sp_invertebase}
\title{dataset6: Documentation of data file - example_sp_invertebase.rda}
\format{
A data frame with 710 rows and 20 variables:
\describe{
\item{source}{invertEbase}
\item{catalogNumber}{CatalogNumber}
\item{basisOfRecord}{type of observations}
\item{occurrenceStatus}{presence or absent}
\item{institutionCode}{Institution code}
\item{verbatimEventDate}{when was this occurrence created}
\item{scientificName}{species name}
\item{individualCount}{abundance}
\item{organismQuantity}{abundance}
\item{abundance}{abundance}
\item{decimalLatitude}{Latitude in decimal degrees}
\item{decimalLongitude}{Longitude in decimal degrees}
\item{coordinateUncertaintyInMeters}{uncertainty of coordiantes}
\item{locality}{location information}
\item{verbatimLocality}{verbatim location information}

\item{municipality}{municipality}
\item{country}{country}
\item{stateProvince}{State or Provinces}
\item{county}{county}
\item{countryCode}{country code}
}
}
\source{
\itemize{
\item this file is downloaded file from invertEBase for species - "Mexacanthina lugubris" and modified field names based on TDWG standard.
}
}
\usage{
example_sp_invertebase
}
\description{
This is a data dump downloaded from invertEbase, as the R package link with InverEbase is currently archive and not maintained, we are providing an example file.
}
\keyword{datasets}
