% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbif_occ.R
\name{gbif_occ}
\alias{gbif_occ}
\title{Extract the data from GBIF for the species and required attributes}
\usage{
gbif_occ(species_name, attribute_list)
}
\arguments{
\item{species_name}{species of interest to get the occurrence data. e.g. "Mexacanthina lugubris"}

\item{attribute_list}{List of attribuites which you want to extract from GBIF. Use the same attribute list to extract data from other sources too.}
}
\value{
a occurrence table from gbif, added and modified few fields
}
\description{
Extract the data from GBIF for the species and required attributes
}
\examples{
species_name <- "Mexacanthina lugubris"
attribute_list <- c("source", "catalogNumber", "basisOfRecord", "occurrenceStatus", "institutionCode", "verbatimEventDate", "scientificName", "individualCount", "organismQuantity", "abundance", "decimalLatitude", "decimalLongitude", "coordinateUncertaintyInMeters", "locality", "verbatimLocality", "municipality", "county", "stateProvince", "country", "countryCode")
gbif_sp_occur <- gbif_occ(species_name, attribute_list)

}
