% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_trail_zero.R
\name{ec_trail_zero}
\alias{ec_trail_zero}
\title{Trail Zeros from the Coordinate Values}
\usage{
ec_trail_zero(coord)
}
\arguments{
\item{coord}{A coordinate value in the numeric format of decimal degree}
}
\value{
A numerical trailed coordinate value.
}
\description{
Trail Zeros from the Coordinate Values
}
\examples{
ec_trail_zero(12.7000000)
ec_trail_zero(45.000000)
}
